make usage of Cisco package repos optional.
This commit adds the variable: $::package_repo that
can be used to specify that the Cisco package repo
should be used to install the openstack packages
on Ubuntu.

The default for this value should default to cisco_repo,
but can also be set to cloud_archive to indicate the Ubuntu
upstream packages should be used.