make compute and controller node matches regexes
This commit updates the node declarations for the
compute and control nodes so that they match on regex.

The motivation behind this is so that I can name
nodes with a unique string (timestamp) so that I can
easily rebuild compute and controller instances without
having to rebuild (or clean up) the build server.

Basically, with a single master, it is hard to re-use
cert names for two reasons:
  - puppet is not very good at reissuing certnames
  - storeconfig entries are not cleaned up (resulting in
  duplicate resources)