Implement multiple Quantum dhcp agents
This commit addresses the following blueprint:

https://blueprints.launchpad.net/openstack-cisco/+spec/multiple-quantum-dhcp-agents

It provides users with the ability to enable multiple
Quantum dhcp agents in their OpenStack cloud.  The default
settting is to run a dhcp agent only on the control node.
This commit adds a commented-out line to a compute node
definition that users can uncomment (and add to additional
compute node definitions) to enable multiple dhcp agents.