Add support for booting specific kernel
Adds support for installing a user-specified kernel package
and setting it to be booted by default.  This satisfies the
ask from the Nexus 1000v team to be able to run a particular
kernel on compute/control nodes so that nodes are known to be
running a kernel that is compatible with Nexus 1000v kernel
modules.

Implements: bp/nagchund