Changed How Load-Balancers are Being Composed
Previously, all load-balancer configuration was contained within
the haproxy-nodes manifest and imported into the site manifest.
This caused a great deal of configuration parameters to be exposed
to end-users.

This change removes haproxy-nodes manifest in favor of the new
openstack-ha::load-balancer class.  This commit relies on the
following commit for the openstack-ha module:

  https://github.com/CiscoSystems/puppet-openstack-ha/commit/cf110249fd71c70100c4c3cd3feb09373d847943