Move $tenant_network_type to provider net section
The $tenant_network_type variable is used to set the segmentation
type for provider networks.  It is also required to be set to
'vlan' when using the Cisco Nexus plugin or when using vlan-based
provider networks.  We previously had this variable in the section
of the file used to configure the Cisco Nexus plugin, but failed
to mention it in the provider networks section.  We've now had a
couple of folks run into problems when attempting to use provider
networks as a result since it wasn't clear that it needed to be set.

This commit moves the variable to the provider networks section,
updates the comments in the provider networks section, and also
providers comments in the Cisco plugin section in order to make it
more clear to users how to use this variable.

Fixes LP bug #1216078