Remove duplicate definition of /var/www
An earlier patch for Bug #1216032 added a /var/www as a resource in
the puppet-coe module and set permissions for it.  However, if the
site.pp file in use doesn't define a $node_gateway, a condition in
core.pp is invoked which also creates /var/www.  The definition in
core.pp is unecessary and is less specific than the one in puppet-coe
and should therefore be removed.

Fixes LP bug #1216032