Restore missing patches from HA merge
During the course of HA developement we had to work out several
merge conflicts.  In the course of doing so, we appear to have
erroneously removed a comment from site.pp that was part of
LP bug #1213230 and re-added a block of code in core.pp that
was removed as part of LP bug #1209373.  These affect only
non-HA deployments.  This patch returns both fixes to their
intended state.

Closes-Bug: #1219077