Add Ceph to all class
The previous commit for this blueprint implemented the first part of
the blueprint's overall goal: it implemented control and compute
functions on one node.  This commit implements a second part: adding
support for Ceph-backed Cinder and Glance.  It also cleans up the
Cisco Nexus parameters for the all class, which were the result of a
commit that was merged at about the same time as part one of the
openstack-all-in-one feature.

Note: getting Ceph working for Cinder and Glance relies on two
upstream commits:

https://review.openstack.org/#/c/47607/
https://review.openstack.org/#/c/47605/

Implements: blueprint openstack-all-in-one