Fix missing rabbit_user and rabbit_host params
The previous rendition of the all class didn't send rabbit_host
and rabbit_user params to the openstack::all class explicitly:
it used defaults.  This works fine for a truly all-in-one standalone
deployment, but causes problems when you try to add additional nodes
to the cloud as they will be configured with a different rabbit user.
This commit rectifies that problem by passing params to the
openstack::all class appropriately.

Implements: blueprint openstack-all-in-one