Add new force_config_drive param to allinone class
The previous version of the allinone class didn't have a parameter
for the nova.conf force_config_drive option because the underlying
openstack::all class didn't support such a parameter.  The
force_config_drive parameter was added to openstack::all in this
change upstream:
https://review.openstack.org/48117

This commit adds a param to the allinone class and passes it through
to the openstack::all class appropriately.

Implements: blueprint openstack-all-in-one