Pass iscsi_ip_address to openstack::all
This commit adds a parameter to the allinone class to allow it to pass
iscsi_ip_address to the openstack::all class.  This allows AIO
deployments to expose an address other than 127.0.0.1 (the default
in openstack::all) which is important if you intend to use an AIO node
as the controller for a multi-node cloud.

Partial-Bug: #1236137