Add Nagios Service Checks to AIO Deployments
The previous version of the allinone class definiton in core.pp
did not include an instantiation of the naginator class that creates
Nagios checks for control-node functions (such as RabbitMQ and
nova-api checks).  Since AIO nodes run control functions, this left
users without any telemetry into controller functions on AIO nodes.
This patch adds a call to the Naginator class to set up the control
checks.

Closes-Bug: #1240089